vehiculares/settings/production.py
import os

from dotenv import load_dotenv
from vehiculares.settings.base import *
from vehiculares.logging import *

load_dotenv(Path.joinpath(BASE_DIR, '.env'))

# SECURITY WARNING: keep the secret key used in production secret!

SECRET_KEY = os.environ.get('SECRET_KEY')

# SECURITY WARNING: don't run with debug turned on in production!
DEBUG = False

ALLOWED_HOSTS = ['*']

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.postgresql_psycopg2',
        'NAME': os.environ.get('DB_NAME'),
        'USER': os.environ.get('DB_USER'),
        'PASSWORD': os.environ.get('DB_PASSWORD'),
        'HOST': os.environ.get('DB_HOST'),
        'PORT': os.environ.get('DB_PORT'),
    }
}

# Application definition

STATIC_ROOT = Path.joinpath(BASE_DIR, 'staticfiles')


STATICFILES_storage = 'whitenoise.storage.CompressedManifestStaticFilesStorage'


