requirements.txt
Django==3.2.10
graphene-django==3.0.0
djangorestframework==3.12.4
django-cors-headers==3.8.0
django-rest-framework-social-oauth2
pytest
pytest-django
graphene-django-extras==0.5.2
drf-yasg==1.20.0
django-filter==2.4.0
graphene-file-upload
django-graphql-auth==0.3.0
python-decouple==3.5
<<<<<<< HEAD
django-health-check==3.8.0
=======
>>>>>>> master
django-extensions==3.1.3
django-graphql-jwt
psycopg2-binary==2.9.8
pillow==10.1.0
<<<<<<< HEAD
=======
django-filter
PyJWT
django-health-check
mixer
coverage



